+++ builtins-test/benches/float_conv.rs
@@ -1,9 +1,9 @@
 #![allow(improper_ctypes)]
 #![cfg_attr(f128_enabled, feature(f128))]
 
+use builtins_test::float_bench;
 use compiler_builtins::float::conv;
 use criterion::{Criterion, criterion_main};
-use testcrate::float_bench;
 
 /* unsigned int -> float */
