+++ builtins-test/benches/float_div.rs
@@ -1,8 +1,8 @@
 #![cfg_attr(f128_enabled, feature(f128))]
 
+use builtins_test::float_bench;
 use compiler_builtins::float::div;
 use criterion::{Criterion, criterion_main};
-use testcrate::float_bench;
 
 float_bench! {
     name: div_f32,
